lms_course_cache folder restructuring
