+++ cms/envs/common.py
@@ -55,6 +55,7 @@ import dealer.git
 from xmodule.modulestore.edit_info import EditInfoMixin
 from xmodule.mixin import LicenseMixin
 
+
 ############################ FEATURE CONFIGURATION #############################
 STUDIO_NAME = "Studio"
 STUDIO_SHORT_NAME = "Studio"
